autoclick/requirements.txt
adbtuils
pillow
